requirements-lock.txt
accelerate==0.29.3
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.3.0
altair-data-server==0.4.1
altair-saver==0.5.0
altair-viewer==0.4.0
annotated-types==0.7.0
anyio==4.10.0
appscript==1.3.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
bleach==6.2.0
bokeh==3.5.1
brotli==1.1.0
cairocffi==1.7.1
cairosvg==2.7.1
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.2
cssselect==1.3.0
cssselect2==0.8.0
cssutils==2.11.1
ctransformers==0.2.27
cycler==0.12.1
datadreamer-dev==0.28.0
dataframe-image==0.2.4
datasets==2.19.0
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
docstring-parser==0.17.0
docx2pdf==0.1.8
et-xmlfile==2.0.0
evaluate==0.4.5
exceptiongroup==1.3.0
ezdxf==1.3.3
faiss-cpu==1.12.0
fastjsonschema==2.21.2
filelock==3.19.1
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2024.3.1
geopandas==1.0.0
graphviz==0.20.3
greenlet==3.0.3
h11==0.16.0
html2image==2.0.7
html5lib==1.1
httpcore==1.0.9
httpx==0.27.2
huggingface-hub==0.23.4
idna==3.10
importlib-metadata==8.7.0
jinja2==3.1.6
joblib==1.5.2
jsonlines==4.0.0
jsonpickle==4.1.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-client==8.6.3
jupyter-core==5.8.1
jupyterlab-pygments==0.3.0
kaleido==0.2.1
kiwisolver==1.4.9
litellm==1.31.14
loguru==0.7.3
lxml==6.0.1
markdown-it-py==4.0.0
markdown2==2.5.0
markupsafe==3.0.2
matplotlib==3.9.0
mdurl==0.1.2
mistune==3.1.4
more-itertools==10.8.0
mpl-finance==0.10.1
mplfinance==0.12.10b0
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
music21==9.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.3
nltk==3.9.1
numpy==1.26.4
openai==1.35.5
openpyxl==3.1.5
optimum==1.27.0
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==25.0
pandas==1.5.3
pandocfilters==1.5.1
pdf2image==1.17.0
peft==0.11.1
pillow==11.3.0
platformdirs==4.4.0
playwright==1.44.0
plotly==5.18.0
portpicker==1.6.0
propcache==0.3.2
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyarrow-hotfix==0.7
pycparser==2.23
pydantic==2.11.9
pydantic-core==2.33.2
pydyf==0.11.0
pyee==11.1.0
pygments==2.19.2
pylatex==1.4.2
pyogrio==0.11.1
pyparsing==3.2.4
pyphen==0.17.2
pyproj==3.7.1
pyro5==5.15
pysocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.1
pytz==2025.2
pyyaml==6.0.2
pyzmq==27.1.0
rdkit==2024.3.5
referencing==0.36.2
regex==2025.9.1
reportlab==4.2.2
requests==2.32.5
rich==14.1.0
ring==0.10.1
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.14.0
seaborn==0.13.2
selenium==4.32.0
sentence-transformers==2.3.0
sentencepiece==0.2.1
serpent==1.41
setfit==1.0.3
setuptools==80.9.0
shapely==2.0.4
shtab==1.7.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
sqlitedict==2.1.0
squarify==0.4.4
svglib==1.5.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tinycss2==1.4.0
tokenizers==0.19.1
toolz==1.0.0
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.40.1
trio==0.31.0
trio-websocket==0.12.2
trl==0.8.1
typeguard==4.4.4
typing-extensions==4.15.0
typing-inspection==0.4.1
tyro==0.9.31
urllib3==2.5.0
vega-datasets==0.9.0
vl-convert-python==1.5.0
weasyprint==62.3
webcolors==24.11.1
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.45.1
wirerope==1.0.0
wsproto==1.2.0
xxhash==3.5.0
xyzservices==2025.4.0
yarl==1.20.1
zipp==3.23.0
zopfli==0.2.3.post1
